+++ bindings/r/tests/testthat/helper-validation.R
@@ -149,8 +149,8 @@ check_stats_lowess <- function(
                     "fraction: %.17g; iterations: %d)"
                 ),
                 max_diff,
-                paste0(sprintf("%.17g", x), collapse = ", "),
-                paste0(sprintf("%.17g", y), collapse = ", "),
+                toString(sprintf("%.17g", x)),
+                toString(sprintf("%.17g", y)),
                 fraction,
                 iterations
             ),